+++ packages/moonstone/VideoPlayer/VideoPlayer.js
@@ -38,6 +38,15 @@ import css from './VideoPlayer.less';
 const SpottableDiv = Spottable('div');
 const Container = SpotlightContainerDecorator({enterTo: ''}, 'div');
 
+// Keycode map for webOS TV
+const keyMap = {
+	'PLAY': 415,
+	'STOP': 413,
+	'PAUSE': 19,
+	'REWIND': 412,
+	'FASTFORWARD': 417
+};
+
 // Video ReadyStates
 // - Commented are currently unused.
 //
@@ -89,6 +98,23 @@ const forwardForwardButtonClick = forward('onForwardButtonClick');
 const forwardJumpBackwardButtonClick = forwardWithPrevent('onJumpBackwardButtonClick');
 const forwardJumpForwardButtonClick = forwardWithPrevent('onJumpForwardButtonClick');
 const forwardPlayButtonClick = forward('onPlayButtonClick');
+
+const AnnounceState = {
+	// Video is loaded but additional announcements have not been made
+	READY: 0,
+
+	// The title should be announced
+	TITLE: 1,
+
+	// The title has been announce
+	TITLE_READ: 2,
+
+	// The infoComponents should be announce
+	INFO: 3,
+
+	// All announcements have been made
+	DONE: 4
+};
 
 /**
  * Every callback sent by [VideoPlayer]{@link moonstone/VideoPlayer} receives a status package,
@@ -518,12 +544,9 @@ const VideoPlayerBase = class extends React.Component {
 		this.moreInProgress = false;	// This only has meaning for the time between clicking "more" and the official state is updated. To get "more" state, only look at the state value.
 		this.prevCommand = (props.noAutoPlay ? 'pause' : 'play');
 		this.speedIndex = 0;
+		this.id = this.generateId();
 		this.titleOffsetCalculated = false;
 		this.selectPlaybackRates('fastForward');
-
-		// localized strings
-		this.playLabel = $L('Play');
-		this.pauseLabel = $L('Pause');
 
 		this.initI18n();
 
@@ -542,6 +565,7 @@ const VideoPlayerBase = class extends React.Component {
 
 		// Re-render-necessary State
 		this.state = {
+			announce: AnnounceState.READY,
 			buffered: 0,
 			currentTime: 0,
 			duration: 0,
@@ -574,6 +598,7 @@ const VideoPlayerBase = class extends React.Component {
 		on('mousemove', this.activityDetected);
 		on('keypress', this.activityDetected);
 		on('keydown', this.handleGlobalKeyDown);
+		on('keyup', this.handleKeyUp);
 		this.attachCustomMediaEvents();
 		this.startDelayedFeedbackHide();
 		this.renderBottomControl.idle();
@@ -633,18 +658,25 @@ const VideoPlayerBase = class extends React.Component {
 		) {
 			this.focusDefaultMediaControl();
 		}
+
+		if (this.state.more !== prevState.more) {
+			this.refocusMoreButton.start();
+		}
 	}
 
 	componentWillUnmount () {
 		off('mousemove', this.activityDetected);
 		off('keypress', this.activityDetected);
 		off('keydown', this.handleGlobalKeyDown);
+		off('keyup', this.handleKeyUp);
 		this.detachCustomMediaEvents();
 		this.stopRewindJob();
 		this.stopAutoCloseTimeout();
 		this.stopDelayedTitleHide();
 		this.stopDelayedFeedbackHide();
+		this.announceJob.stop();
 		this.renderBottomControl.stop();
+		this.refocusMoreButton.stop();
 	}
 
 	//
@@ -724,14 +756,26 @@ const VideoPlayerBase = class extends React.Component {
 		this.autoCloseJob.stop();
 	}
 
-	showControls = () => {
-		// Read the title
-		this.announce(this.props.title);
+	generateId = () => {
+		return Math.random().toString(36).substr(2, 8);
+	}
 
+	showControls = () => {
 		this.startDelayedFeedbackHide();
 		this.startDelayedTitleHide();
 		forwardControlsAvailable({available: true}, this.props);
+
+		let {announce} = this.state;
+		if (announce === AnnounceState.READY) {
+			// if we haven't read the title yet, do so this time
+			announce = AnnounceState.TITLE;
+		} else if (announce === AnnounceState.TITLE) {
+			// if we have read the title, advance to INFO so title isn't read again
+			announce = AnnounceState.TITLE_READ;
+		}
+
 		this.setState({
+			announce,
 			bottomControlsRendered: true,
 			bottomControlsVisible: true,
 			feedbackVisible: true,
@@ -747,6 +791,13 @@ const VideoPlayerBase = class extends React.Component {
 	}
 
 	autoCloseJob = new Job(this.hideControls)
+
+	refocusMoreButton = new Job(() => {
+		// Readout 'more' or 'back' button explicitly.
+		let selectedButton = Spotlight.getCurrent();
+		selectedButton.blur();
+		selectedButton.focus();
+	}, 100)
 
 	startDelayedTitleHide = () => {
 		if (this.props.titleHideDelay) {
@@ -797,6 +848,33 @@ const VideoPlayerBase = class extends React.Component {
 	showControlsFromPointer = () => {
 		Spotlight.setPointerMode(false);
 		this.showControls();
+	}
+
+	handleKeyUp = (ev) => {
+		const {PLAY, PAUSE, STOP, REWIND, FASTFORWARD} = keyMap;
+
+		switch (ev.keyCode) {
+			case PLAY:
+				this.play();
+				break;
+			case PAUSE:
+				this.pause();
+				break;
+			case REWIND:
+				if (!this.props.noRateButtons) {
+					this.rewind();
+				}
+				break;
+			case FASTFORWARD:
+				if (!this.props.noRateButtons) {
+					this.fastForward();
+				}
+				break;
+			case STOP:
+				this.pause();
+				this.seek(0);
+				break;
+		}
 	}
 
 	handleGlobalKeyDown = this.handle(
@@ -874,6 +952,9 @@ const VideoPlayerBase = class extends React.Component {
 	reloadVideo = () => {
 		// When changing a HTML5 video, you have to reload it.
 		this.video.load();
+		this.setState({
+			announce: AnnounceState.READY
+		});
 	}
 
 	/**
@@ -902,6 +983,7 @@ const VideoPlayerBase = class extends React.Component {
 		this.setPlaybackRate(1);
 		this.send('play');
 		this.prevCommand = 'play';
+		this.announce($L('Play'));
 	}
 
 	/**
@@ -916,6 +998,7 @@ const VideoPlayerBase = class extends React.Component {
 		this.setPlaybackRate(1);
 		this.send('pause');
 		this.prevCommand = 'pause';
+		this.announce($L('Pause'));
 	}
 
 	/**
@@ -1263,10 +1346,8 @@ const VideoPlayerBase = class extends React.Component {
 		forwardPlayButtonClick(ev, this.props);
 		if (this.state.paused) {
 			this.play();
-			this.announce(this.playLabel);
 		} else {
 			this.pause();
-			this.announce(this.pauseLabel);
 		}
 	}
 	onForward = (ev) => {
@@ -1290,9 +1371,11 @@ const VideoPlayerBase = class extends React.Component {
 			// Interrupt the title-hide since we don't want it hiding autonomously in "more".
 			this.stopDelayedTitleHide();
 		}
+
 		this.setState({
 			more: !this.state.more,
-			titleVisible: true
+			titleVisible: true,
+			announce: this.state.announce < AnnounceState.INFO ? AnnounceState.INFO : AnnounceState.DONE
 		});
 	}
 
@@ -1317,6 +1400,24 @@ const VideoPlayerBase = class extends React.Component {
 	renderBottomControl = new Job(() => {
 		this.setState({bottomControlsRendered: true});
 	});
+
+	getControlsAriaProps () {
+		if (this.state.announce === AnnounceState.TITLE) {
+			return {
+				role: 'alert',
+				'aria-live': 'off',
+				'aria-labelledby': `${this.id}_title`
+			};
+		} else if (this.state.announce === AnnounceState.INFO) {
+			return {
+				role: 'alert',
+				'aria-live': 'off',
+				'aria-labelledby': `${this.id}_info`
+			};
+		}
+
+		return null;
+	}
 
 	render () {
 		const {backwardIcon, children, className, forwardIcon, infoComponents, jumpBackwardIcon, jumpButtonsDisabled, jumpForwardIcon, leftComponents, noAutoPlay, noJumpButtons, noRateButtons, noSlider, pauseIcon, playIcon, rateButtonsDisabled, rightComponents, source, style, thumbnailSrc, title, ...rest} = this.props;
@@ -1344,6 +1445,7 @@ const VideoPlayerBase = class extends React.Component {
 
 		// Handle some cases when the "more" button is pressed
 		const moreDisabled = !(this.state.more);
+		const controlsAriaProps = this.getControlsAriaProps();
 
 		return (
 			<div className={css.videoPlayer + (className ? ' ' + className : '')} style={style} onClick={this.activityDetected} onKeyDown={this.activityDetected} ref={this.setPlayerRef}>
@@ -1364,14 +1466,15 @@ const VideoPlayerBase = class extends React.Component {
 				</Overlay>
 
 				{this.state.bottomControlsRendered ?
-					<div className={css.fullscreen + ' enyo-fit scrim'} style={{display: this.state.bottomControlsVisible ? 'block' : 'none'}}>
+					<div className={css.fullscreen + ' enyo-fit scrim'} style={{display: this.state.bottomControlsVisible ? 'block' : 'none'}} {...controlsAriaProps}>
 						<Container className={css.bottom} data-container-disabled={!this.state.bottomControlsVisible}>
 							{/* Info Section: Title, Description, Times */}
 							<div className={css.infoFrame}>
 								<MediaTitle
+									id={this.id}
+									infoVisible={this.state.more}
 									title={title}
 									visible={this.state.titleVisible}
-									infoVisible={this.state.more}
 								>
 									{infoComponents}
 								</MediaTitle>
@@ -1411,7 +1514,6 @@ const VideoPlayerBase = class extends React.Component {
 								noJumpButtons={noJumpButtons}
 								noRateButtons={noRateButtons}
 								onBackwardButtonClick={this.onBackward}
-								onClick={this.resetAutoTimeout}
 								onForwardButtonClick={this.onForward}
 								onJumpBackwardButtonClick={this.onJumpBackward}
 								onJumpForwardButtonClick={this.onJumpForward}
@@ -1420,9 +1522,9 @@ const VideoPlayerBase = class extends React.Component {
 								onToggleMore={this.onMoreClick}
 								paused={this.state.paused}
 								pauseIcon={pauseIcon}
-								pauseLabel={this.pauseLabel}
+								pauseLabel={$L('Pause')}
 								playIcon={playIcon}
-								playLabel={this.playLabel}
+								playLabel={$L('Play')}
 								rateButtonsDisabled={rateButtonsDisabled}
 								rightComponents={rightComponents}
 								showMoreComponents={this.state.more}