+++ packages/moonstone/CHANGELOG.md
@@ -10,7 +10,14 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Changed
 
+- `moonstone/VideoPlayer` to handle play, pause, stop, fast forward and rewind on remote controller
+
 ### Fixed
+
+- `moonstone/Scroller` not scrolled via 5 way when `moonstone/ExpandableList` is opened.
+- `moonstone/VirtualList` not to let the focus move outside of container even if there are children left when navigating with 5way
+- `moonstone/Scrollable` to update disability of paging controls when the scrollbar is set to `visible` and the content becomes shorter.
+- `moonstone/VideoPlayer` to focus on hover over play/pause button when video is loading.
 
 ## [1.5.0] - 2017-07-19
 
@@ -33,6 +40,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/ExpandableItem`, `moonstone/ExpandableList`, `moonstone/ExpandablePicker`, `moonstone/DatePicker`, and `moonstone/TimePicker` to pause spotlight when animating in 5-way mode
 - `moonstone/Spinner` to position the text content under the spinner, rather than to the right side
 - `moonstone/VideoPlayer` to include hour when announcing the time while scrubbing
+- `moonstone/GridListImageItem` to require a `source` prop and not have a default value
 
 ### Fixed